+++ src/Generator/GeneratorHookRunner.php
@@ -3,8 +3,8 @@
 namespace Helmich\Schema2Class\Generator;
 
 use Laminas\Code\Generator\ClassGenerator;
-use Laminas\Code\Generator\EnumGenerator\EnumGenerator;
 use Laminas\Code\Generator\FileGenerator;
+use Helmich\Schema2Class\Generator\PhpParserEnumGenerator;
 
 trait GeneratorHookRunner
 {
@@ -38,7 +38,7 @@ trait GeneratorHookRunner
         }
     }
 
-    public function onEnumCreated(string $enumName, EnumGenerator $enum): void
+    public function onEnumCreated(string $enumName, PhpParserEnumGenerator $enum): void
     {
         foreach ($this->hooks as ['hook' => $hook]) {
             if ($hook instanceof Hook\EnumCreatedHook) {